Keep MAVLink/TCP connection alive
Before this commit, the TCP connection to the Xbee was shutdown before entering autonmous mode. This seemed to lock up the device when more than one xbee was on the network. The new logic does not shutdown the connection and just prevents sending data over the TCP connection in autonomous mode.